Generator complete for linux; needs configuration

+++ file/file_linux.odin
@@ -6,7 +6,7 @@ exists :: proc(path: string) -> bool
 {
     if stat, err := os.stat(path); err == os.ERROR_NONE 
     {
-        return os.S_ISREG(stat.mode) || os.S_ISDIR(stat.mode);
+        return os.S_ISREG(u32(stat.mode)) || os.S_ISDIR(u32(stat.mode));
     }
     return false;
 }
